# openapi.yaml
openapi: 3.0.3
info:
  title: Cling AI Go 用户端 API
  version: 0.1.0-local
  description: 本文档只描述本地 Go 用户端合同；管理后台、Animate、真实支付供应商和生产地址不在范围内。
servers:
  - url: http://127.0.0.1:18000
    description: 本地 Gateway
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      summary: 注册并创建 Go 会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '200': { $ref: '#/components/responses/LoginResponse' }
  /api/auth/login:
    post:
      summary: 密码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/LoginResponse' }
...
  /api/auth/bind:
    post:
      summary: 游客升级为已绑定用户
      description: 服务端验证外部凭据并升级同一个用户，不创建第二个账号。
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object, required: [provider, credential], properties: { provider: { type: string }, credential: { type: string, writeOnly: true } } } }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
...
  /api/media/images:
    post:
      summary: 上传用户图片素材
      requestBody:
        required: true
        content:
          multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } }
      responses:
        '201': { $ref: '#/components/responses/CreationResponse' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, timezone]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, writeOnly: true }
        timezone: { type: string }
        displayName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, writeOnly: true }
...
  responses:
    LoginResponse: { description: 会话结果, content: { application/json: { schema: { type: object } } } }
    UserResponse: { description: 用户结果, content: { application/json: { schema: { type: object } } } }
    CreationResponse: { description: 创建结果, content: { application/json: { schema: { type: object } } } }
...